searching_chatbot/chatbot-webapp/requirements.txt
Flask>=2.2
python-dotenv>=1.0
SQLAlchemy>=1.4
# groq is optional; include if you want Groq integration
groq>=0.9.0
